requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
jinja2==3.1.4
itsdangerous==2.2.0

# Async HTTP
httpx==0.28.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Auth
authlib==1.4.0
pyjwt[crypto]==2.10.1

# Document processing
python-docx==1.1.2

# Utilities
pydantic==2.10.3
pydantic-settings==2.7.0
structlog==24.4.0
tenacity==9.0.0
python-dotenv==1.0.1
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx  # already listed, used by test client
